Cleaning Set = Clean slate
