+++ globals.py
@@ -40,7 +40,11 @@ avSetDistance = 600
 # increasing: Means that the distance will increase steadily from one block to the other
 # decreasing: Means that the distance will decrease steadily from one block to the other
 # pyramid: Means that the distance will increase and then decrease. 
-variationTypes = ["equal", "increasing", "decreasing", "pyramid", "distanceRep"]
+setTypes = ["equalDistance",
+            "increasingDistance",
+            "decreasingDistance",
+            "pyramidDistance"]
+#              "distanceRep"]
 
 # Minimal difference distance between two blocks (m)
 # If this variable is set at 25m, this means that we could have a set where the blocks are for example 25m, 50m, 75m, etc. If set at 50m, then the
